% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_19_data.R
\docType{data}
\name{covid_19_data}
\alias{covid_19_data}
\title{Global epidemic situation of COVID-19}
\format{
A data frame (tibble) with 116355 observations and 4 variables:
\itemize{
\item \strong{observation_date}: Observation date
\item \emph{country_region}*: The name of the country or region
\item \strong{type}: The type of the case (confirmed, recovered, deaths)
\item \strong{count}: The number of daily cases of each type
}
}
\source{
This data comes from \link{World Health Organization}\url{https://www.who.int/emergencies/diseases/novel-coronavirus-2019}
}
\usage{
covid_19_data
}
\description{
This dataset contains daily epidemic information of COVID-19 for 222 countries or regions, from 2020-01-22 to 2020-09-24.
}
\examples{
library(COVID19bb)
data(covid_19_data)

}
\keyword{datasets}
