% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_output.R
\name{distribution_map}
\alias{distribution_map}
\title{Create distribution map}
\usage{
distribution_map(data, Var1, Var2)
}
\arguments{
\item{data:}{The data used to draw the distribution
Var1: The variable whose distribution is intended to be displayed in the map
Var2: The variable used to join the data with worldgeojson list. (i.e. the country or the region name)}
}
\value{
A distribution world map of certain objects
}
\description{
Function \code{distribution_map()} can create a distribution map of desired objects
}
\examples{
\dontrun{
distribution_map(data = covid_19_data,Var1 = count, Var2 = country_region)
}

}
