% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Load the shiny app}
\usage{
launch_app()
}
\description{
This function is used to launch the COVID-19 Global Analysis Shiny Application from the \code{COVID19bb} package.
}
\details{
This app is intended to be used as an intuitive and easy-to-understand visualization tool for COVID-19 research, users can observe the total distribution of confirmed cases, death cases and recovered cases around the world by selecting the type of the distribution they want to view, as well as exploring the trends of mortality, diagnosis recovered counts among provinces or states of different countries all over the world.
}
\examples{
\dontrun{
launch_app()
}

}
