% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_input.R
\name{select_input}
\alias{select_input}
\title{Add selectInput sidebar for shiny app}
\usage{
select_input(inputId, label, choices, selected)
}
\arguments{
\item{inputId:}{The ID of the input object
label: The text label set for the input object
choices: List of values to select from
selected: The default selected option}
}
\value{
A shiny selected-input sidebar
}
\description{
Function \code{select_input()} can help users select input object within the Shiny Application
}
\examples{
\dontrun{
select_input("type",label = NULL, choices = NULL,selected = NULL)
}

}
