% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_filter.R
\name{table_filter}
\alias{table_filter}
\title{Table after filtering of daily cases of COVID-19}
\usage{
table_filter(data, country, type, date)
}
\arguments{
\item{data}{The data used to create the table}

\item{country}{The country that you want to exlpore}

\item{type}{The type of the case}

\item{date}{THe date used to observe}
}
\value{
A table after filtering of daily cases of COVID-19
}
\description{
Function \code{table_filter()} can create a table after filtering of daily cases of COVID-19
}
\examples{
\dontrun{
table_filter(covid_19_data,"China","deaths","2020-08-23")
}

}
