% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldgeojson.R
\docType{data}
\name{worldgeojson}
\alias{worldgeojson}
\title{worldgeojson}
\format{
A list in geojson format
}
\source{
This data comes from \link{highcharter v0.8.2}<https://code.highcharts.com/mapdata/custom/world.js >
}
\usage{
worldgeojson
}
\description{
This dataset is the world map in geojson format (list)
}
\examples{
library(COVID19bb)
data(covid_19_all)

}
\keyword{datasets}
